refactor(ShortestPathWithDijkstra.java) : 제네릭 타입 명시
