+++ src/main/java/wooteco/subway/path/application/PathService.java
@@ -5,7 +5,7 @@ import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.line.domain.Section;
 import wooteco.subway.path.domain.Path;
 import wooteco.subway.path.dto.PathResponse;
-import wooteco.subway.path.infrastructure.ShortestPath;
+import wooteco.subway.path.domain.ShortestPath;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 
@@ -26,7 +26,6 @@ public class PathService {
 
     public PathResponse findPaths(long sourceId, long targetId) {
         List<Section> sections = sectionDao.findAll();
-
         Station source = stationDao.findById(sourceId);
         Station target = stationDao.findById(targetId);
         Path path = shortestPath.getPath(source, target, sections);
