+++ src/main/java/wooteco/subway/path/infrastructure/ShortestPathWithDijkstra.java
@@ -7,6 +7,7 @@ import org.jgrapht.graph.WeightedMultigraph;
 import org.springframework.stereotype.Component;
 import wooteco.subway.line.domain.Section;
 import wooteco.subway.path.domain.Path;
+import wooteco.subway.path.domain.ShortestPath;
 import wooteco.subway.station.domain.Station;
 
 import java.util.List;
@@ -15,7 +16,7 @@ import java.util.List;
 public class ShortestPathWithDijkstra implements ShortestPath {
     @Override
     public Path getPath(Station source, Station target, List<Section> sections) {
-        WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph(DefaultWeightedEdge.class);
+        WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph<>(DefaultWeightedEdge.class);
 
         for (Section section : sections) {
             graph.addVertex(section.getUpStation());
@@ -23,8 +24,8 @@ public class ShortestPathWithDijkstra implements ShortestPath {
             graph.setEdgeWeight(graph.addEdge(section.getUpStation(), section.getDownStation()), section.getDistance());
         }
 
-        DijkstraShortestPath dijkstraShortestPath = new DijkstraShortestPath(graph);
-        GraphPath path = dijkstraShortestPath.getPath(source, target);
+        DijkstraShortestPath<Station, DefaultWeightedEdge> dijkstraShortestPath = new DijkstraShortestPath<>(graph);
+        GraphPath<Station, DefaultWeightedEdge> path = dijkstraShortestPath.getPath(source, target);
         return new Path(path.getVertexList(), (int) path.getWeight());
     }
 }
